refactor(composio): update provider module references
Point Composio and task source integrations to the provider module under integrations instead of memory sync. This preserves existing behavior while matching the provider module's new location.

Auto-committed-on: macbook
