+++ src/openhuman/integrations/task_sources/ops.rs
@@ -10,7 +10,7 @@ use std::sync::Arc;
 use serde_json::{json, Value};
 
 use crate::openhuman::config::Config;
-use crate::openhuman::memory::sync::composio::providers::{
+use crate::openhuman::integrations::composio::providers::{
     get_provider, NormalizedTask, ProviderContext, TaskContainer,
 };
 use crate::rpc::RpcOutcome;
