+++ src/openhuman/integrations/task_sources/pipeline_tests.rs
@@ -2,7 +2,7 @@ use super::*;
 use crate::openhuman::config::Config;
 use crate::openhuman::integrations::task_sources::store;
 use crate::openhuman::integrations::task_sources::types::{FilterSpec, ProviderSlug, SourceTarget};
-use crate::openhuman::memory::sync::composio::providers::{
+use crate::openhuman::integrations::composio::providers::{
     register_provider, ComposioProvider, NormalizedTask, ProviderContext, ProviderUserProfile,
     TaskFetchFilter,
 };
