+++ src/openhuman/integrations/task_sources/types.rs
@@ -64,7 +64,7 @@ pub enum FilterSpec {
         /// How to fetch: Composio connection, local `gh`/REST, or `auto`
         /// (Composio-first with local fallback). Defaults to `auto`.
         #[serde(default)]
-        fetch_mode: crate::openhuman::memory::sync::composio::providers::GithubFetchMode,
+        fetch_mode: crate::openhuman::integrations::composio::providers::GithubFetchMode,
         #[serde(default)]
         extra: Value,
     },
