+++ src/openhuman/memory/read_rpc/admin.rs
@@ -2,7 +2,7 @@ use anyhow::{Context, Result};
 
 use crate::openhuman::config::Config;
 use crate::openhuman::memory::api::provider::ForgetSelector;
-use crate::openhuman::memory::sync::composio::providers::sync_state::KV_NAMESPACE;
+use crate::openhuman::integrations::composio::providers::sync_state::KV_NAMESPACE;
 use crate::rpc::RpcOutcome;
 use tinymemory_api::chunks::SourceKind;
